HDDS-9087. Ozone Recon Heatmap code refactoring.
